requirements.txt
openai
aiohttp
aiosignal
waitress
async-timeout 
requests 
importlib-metadata 
MarkupSafe
multidict
tqdm
blinker
click
colorama
Flask
gunicorn
itsdangerous
Jinja2
Werkzeug
python-dotenv